requirements.txt
-e .
argparse==1.2.*
beautifulsoup4>=4.8,<5.0
cairosvg
cssutils==1.0.*
google-api-python-client==1.*
google-auth
google-auth-oauthlib
Jinja2==2.8.*
mock
oauthlib
premailer
pygal
python-dateutil
requests==2.10.*
simplejson==3.6.*
tendo==0.2.*
